- minor refactoring: make block-reencoder use the huffman encoder from support directory - several bug fixes - some support for static huffman and stored blocks - some support for empty blocks at the end of the file number of known, working (=tested) deflate streams rises fromo 13 to 18

+++ preflate_hash_chain.h
@@ -20,7 +20,7 @@
 
 struct PreflateHashIterator {
   const unsigned short* chain;
-  const unsigned short* chainDepth;
+  const unsigned * chainDepth;
   const unsigned refPos;
   const unsigned maxDist;
   unsigned curPos, curDist;
@@ -28,7 +28,7 @@ struct PreflateHashIterator {
 
   PreflateHashIterator(
       const unsigned short* chain_,
-      const unsigned short* depth_,
+      const unsigned * depth_,
       const unsigned refPos_,
       const unsigned maxDist_,
       unsigned startPos_)
@@ -67,7 +67,7 @@ struct PreflateHashIterator {
 struct PreflateHashChainExt {
   PreflateInput _input;
   unsigned short* head;
-  unsigned short* chainDepth;
+  unsigned * chainDepth;
   unsigned short* prev;
   unsigned char hashBits, hashShift;
   unsigned short runningHash, hashMask;
